Deltamatrix {npsurv}	R Documentation

Delta matrix

Description

Deltamatrix computes the Delta matrix, along with maximal intersection intervals, for a set of intervals.

Usage

Deltamatrix(LR) 

Arguments

LR
two-column matrix, each row of which stores an censoring interval of the form (Li, Ri]. If Li = Ri, it is an exact observation.

Details

An intersection interval is a nonempty intersection of any combination of the given intervals, and a maximal intersection interval is an intersection interval that contains no other intersection interval.

The Delta matrix is a matrix of indicators (TRUE or FALSE). The rows correspond to the given interval-censored observations, and the columns the maximal intersection intervals. A TRUE value of the (i,j)-th element means that the i-th observation covers the j-th maximal intersection interval, and a FALSE value means the opposite.

Value

A list with components:

left
left endpoints of the maximal intersection intervals.

right
right endpoints of the maximal intersection intervals.

Delta
logical matrix, for the Delta matrix.

Author(s)

Yong Wang <yongwang@auckland.ac.nz>

References

Wang, Y. (2008). Dimension-reduced nonparametric maximum likelihood computation for interval-censored data. Computational Statistics & Data Analysis, 52, 2388-2402.

See Also

icendata, idf.

Examples

(x = cbind(1:5,1:5*3-2))
Deltamatrix(x)

[Package npsurv version 0.4-0.1 Index]