acfail {npsurv}	R Documentation

Air Conditioner Failure Data

Description

Contains the number of operating hours between successive failure times of the air conditioning systems in Boeing airplanes

Usage

acfail

Format

A numeric vector storing the failure times.

Source

Proschan (1963)

References

Proschan, F. (1963). Theoretical explanation of observed decreasing failure rate. Technometrics, 5, 375-383.

See Also

Uhaz.

Examples

data(acfail)
r = Uhaz(acfail, deg=2)
plot(r$h, fn="h")
plot(r$h, fn="d")

[Package npsurv version 0.4-0.1 Index]