cancer {npsurv}	R Documentation

Breast Retraction Times after Beast Cancer Treatments.

Description

Contains the breast retraction times in months for 94 breast cancer patients who received either radiation therapy or radiation therapy plus adjuvant chemotherapy.

Usage

cancer

Format

A data frame with 94 observations and 3 variables:

L: left-end points of the interval-censored retraction times;

R: right-end points of the interval-censored retraction times;

group: either RT (radiation therapy) or RCT (radiation therapy plus adjuvant chemotherapy).

Source

Finkelstein and Wolfe (1985).

References

Finkelstein, D. M. and R. A. Wolfe (1985). A semiparametric model for regression analysis of interval-censored failure time data. Biometrics, 41, pp.933-945.

See Also

npsurv.

Examples

data(cancer)
i = cancer$group == "RT"
plot(npsurv(cancer[i,1:2]), xlim=c(0,60))
plot(npsurv(cancer[!i,1:2]), add=TRUE, col="green3")

[Package npsurv version 0.4-0.1 Index]