marijuana {npsurv}	R Documentation

Angina Pectoris Survival Data

Description

Contains the answers of 191 California high school students to the question: "When did you first use marijuana?". An answer can be an exact age, or "I have never used it", which gives rise to a right-censored observation, or "I have used it but cannot recall just when the first time was", which gives rise to a left-censored observation.

Usage

marijuana

Format

A data frame with 21 observations and 3 variables:

L: left-end point of an interval-censored time;

R: right-end point of an interval-censored time;

count: number of students in the interval.

Source

Turnbull and Weiss (1978). See also Klein and Moeschberger (1997), page 17.

References

Turnbull and Weiss (1978). A likelihood ratio statistic fortesting goodness of fit with randomly censored data. Biometrics, 34, 367-375.

Klein and Moeschberger (2003). Survival Analysis: Techniques for Censored and Truncated Data (2nd ed.). Springer

See Also

npsurv.

Examples

data(marijuana)
r = Uhaz(marijuana, deg=2)
plot(r$h, fn="h")
plot(r$h, fn="s")

[Package npsurv version 0.4-0.1 Index]